requirements-new.txt
python-telegram-bot==21.0
httpx==0.27.0
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1

